Added Edit Functionality to Comments
Comments now have the ability to be edited.  View file for Comments has
been edited to offer editing, routes have been added to accommodate the
edit ability, and the controller has 2 methods added to make this
possible.  Changes to the controller methods are the addition of the
'editcomment' and 'updatecomment' methods.  The 'editcomment' method
serves as a placeholder for the action call to redirect to the
'editcomment' view file.  The 'updatecomment'  method contains the code
to find the record being edited and saving the changed comment.  It also
redirects back to the review that the comment was edited in when
submitted as well as having a back link to cancel out of the edit.