Search Bar Rounded and Edit/Delete Link fix
The search bar has been rounded in the application.html.erb view for the
dashboard overlay.  Also the edit and delete links for posts are only
available if the user was the one that posted the review or comment.
The 'reviews_controller.rb' file has been modified with a method to
check for whether the current user signed in is the same user that
posted the review or comment.  If that user matches, then the
edit/delete options will be available.  Future update will allow for
admin or moderator level members to edit/delete posts as well.