Added Avatars, Major Restructure of Application
Avatars were added as a table, as well as a controller and series of
views for MVC.  Routing table updated.  'application.html.erb' file was
updated to accommodate for additional tables.  Paperclip GEM feature
added to Avatars Model.  Directory was also created to store images
uploaded by users for their avatar.  The reviews_controller had some
methods that were related to the dashboard frame moved out to the
'application_controller.rb' file for cleanup of code.